fix: moved vidazoo interpretResponse to lib
